[#14]refactor(moderation):report count only stay(#34)
