+++ src/main/java/com/matjom/matjom/feed/entity/review/ReviewStatus.java
@@ -3,4 +3,12 @@ package com.matjom.matjom.feed.entity.review;
 public enum ReviewStatus {
     ACTIVE, // 9월 26일 최종: 최소 상태만 유지
     DELETED // 9월 26일 최종: 숨김 상태 제거
+
+    //DELETED 베이스 엔티티에 있는거 사용해서 삭제 여부 알 수있다.
 }
+
+//코드에 주석으로 코드 기능  및 목적 정리 필요
+//swagger api
+//api test 스크린샷- PR로 합찬 후
+//노션이나 md파일로 디렉토리 구조 및 각 파일이 어디에 사용되는지 정리하기
+//각 테스트코드에서 when given then 에 대한 로직이 이해할 수 있도록 주석을 달아서 팀원들이 이해할 수 있도록
